Move common.rs into ratelimiter::ChorusRequest
