--- src/api/common.rs
@@ -1,57 +1,0 @@
-use reqwest::RequestBuilder;
-use serde::Deserialize;
-use serde_json::from_str;
-
-use crate::{
-    errors::{ChorusError, ChorusResult},
-    instance::UserMeta,
-    ratelimiter::LimitedRequester,
-};
-
-use super::limits::LimitType;
-
-/// Sends a request to wherever it needs to go. Returns [`Ok(())`] on success and
-/// [`Err(ChorusLibError)`] on failure.
-pub async fn handle_request_as_result(
-    request: RequestBuilder,
-    user: &mut UserMeta,
-    limit_type: LimitType,
-) -> ChorusResult<()> {
-    match handle_request(request, user, limit_type).await {
-        Ok(_) => Ok(()),
-        Err(e) => Err(ChorusError::InvalidResponseError {
-            error: e.to_string(),
-        }),
-    }
-}
-
-pub async fn deserialize_response<T: for<'a> Deserialize<'a>>(
-    request: RequestBuilder,
-    user: &mut UserMeta,
-    limit_type: LimitType,
-) -> ChorusResult<T> {
-    let response = handle_request(request, user, limit_type).await.unwrap();
-    let response_text = match response.text().await {
-        Ok(string) => string,
-        Err(e) => {
-            return Err(ChorusError::InvalidResponseError {
-                error: format!(
-                    "Error while trying to process the HTTP response into a String: {}",
-                    e
-                ),
-            });
-        }
-    };
-    let object = match from_str::<T>(&response_text) {
-        Ok(object) => object,
-        Err(e) => {
-            return Err(ChorusError::InvalidResponseError {
-                error: format!(
-                    "Error while trying to deserialize the JSON response into T: {}",
-                    e
-                ),
-            })
-        }
-    };
-    Ok(object)
-}
